pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "icd-collector"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "streamlit>=1.35",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.4"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
